first pass, more complicated because no docker-hub image to start from
