--- assets/arm/assets.txt
@@ -1,6 +1,0 @@
-busybox 1549995718 f1382a9b09caa5357b15e52b1701783d
-libdisableselinux.so 1549995718 c3e3d58f0d7ea5484134f24757df92c3
-rootfs.tar.gz.part00 1549995718 8e3911c0d9636533614e0a8f0fb299b4
-rootfs.tar.gz.part01 1549995719 e43ee22607a34a40993cba7e3c717e0a
-rootfs.tar.gz.part02 1549995719 361e257b54f8706ab2c31c44ee0df200
-rootfs.tar.gz.part03 1549995719 d41d8cd98f00b204e9800998ecf8427e
