--- assets/arm64/assets.txt
@@ -1,6 +1,0 @@
-busybox 1549995719 4cf786269144335c65d13c73d7e7d4be
-libdisableselinux.so 1549995719 336721a501924233d7a06877adb67695
-rootfs.tar.gz.part00 1549995719 a0d1c948d8954dc5e0669723ef272a5d
-rootfs.tar.gz.part01 1549995719 a8048588b404e6eae34fe5ce5e4c1b71
-rootfs.tar.gz.part02 1549995720 1cd7fb90e18af4c18ebf23e20f249d47
-rootfs.tar.gz.part03 1549995720 d41d8cd98f00b204e9800998ecf8427e
