--- assets/x86/assets.txt
@@ -1,6 +1,0 @@
-busybox 1549995720 866a86a0eadca237e011acdd33027792
-libdisableselinux.so 1549995720 a18df06adaf4367addbec0ebe9d0b84b
-rootfs.tar.gz.part00 1549995720 dab586097377815b1213146ed3ec5b52
-rootfs.tar.gz.part01 1549995721 a42fa8982ef63c172ea1ea5f865bfc88
-rootfs.tar.gz.part02 1549995721 d176f7cb1976b90c8a137b124dd78589
-rootfs.tar.gz.part03 1549995721 d41d8cd98f00b204e9800998ecf8427e
