--- assets/x86_64/assets.txt
@@ -1,6 +1,0 @@
-busybox 1549995721 47f66a428235dd5f51c623c8a4cc96df
-libdisableselinux.so 1549995721 ea11594aee1343f6111aca59561dcd54
-rootfs.tar.gz.part00 1549995721 b85ee66e5d9dfa866e9409115b554a7a
-rootfs.tar.gz.part01 1549995721 8b965e75660193845370f82fbb5fd0a6
-rootfs.tar.gz.part02 1549995721 b68a387c5297b199868ceafe8ec83aa9
-rootfs.tar.gz.part03 1549995721 d41d8cd98f00b204e9800998ecf8427e
