+++ build_image.yml
@@ -1,0 +1,12 @@
+version: '3.2'
+
+services:
+   imagebuild:
+     build: 
+       context: .
+       dockerfile: Dockerfile.build_image
+       args:
+         RUN_CMD: input/build_image.sh ${IMAGE_ARCH}
+     volumes:
+       - ./output:/output
+       - ./input:/input
