--- scripts/addNonRootUser.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-useradd user -s /bin/bash -m -u 2048
-echo user:userland | chpasswd
-chsh -s /bin/bash user
