--- scripts/shrinkRootfs.sh
@@ -1,3 +1,0 @@
-#! /bin/bash
-
-apt-get clean
