+++ x86_64.yml
@@ -1,0 +1,8 @@
+version: '3.2'
+
+services:
+   imagebuild:
+     build: 
+       args:
+         IMAGE_ARCH: amd64
+         QEMU_FILE: qemu-x86_64-static
